requirements.txt
audeer==2.2.0
audiofile==1.5.0
numpy==2.1.0
pandas==2.2.2
omegaconf==2.3.0
audformat==1.3.0
auglib==1.0.4
torch==2.1.0
audmetric==1.2.2
transformers==4.44.2
datasets==2.21.0
audb==1.9.0
hydra-core==1.3.2

matplotlib==3.9.2
audinterface==1.2.2
audplot==1.4.7
PyYAML==6.0.2
seaborn==0.13.2
scipy==1.14.1